[tests] Move `regression.py` to the Python package
As a result, the regression tests can be run with just `python3 -m
pytest`, i.e. without an additional wrapper that overrides `PYTHONPATH`.
