+++ LibOS/shim/test/regression/test_libos.py
@@ -7,7 +7,7 @@ import signal
 import subprocess
 import unittest
 
-from regression import (
+from graminelibos.regression import (
     HAS_SGX,
     ON_X86,
     RegressionTestCase,
